+++ opendevin/server/agent/agent.py
@@ -15,6 +15,7 @@ from opendevin.events.observation import (
     NullObservation,
 )
 from opendevin.events.serialization.action import action_from_dict
+from opendevin.events.serialization.event import event_to_dict
 from opendevin.events.stream import EventSource, EventStream, EventStreamSubscriber
 from opendevin.llm.llm import LLM
 from opendevin.server.session import session_manager
@@ -139,7 +140,7 @@ class AgentUnit:
         if isinstance(event, NullObservation):
             return
         if event.source == 'agent':
-            await self.send(event.to_dict())
+            await self.send(event_to_dict(event))
             return
 
     def close(self):
